+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Switched to calendar versioning.
 * Added support for reactions in both directions, including backfill.
+* Added support for replies in both directions, including backfill.
 
 # v0.1.2 (2025-07-16)